+++ src/main/java/apiquality/sonar/asyncapi/checks/schemas/AAR046AvroRecordDocCheck.java
@@ -1,0 +1,18 @@
+package apiquality.sonar.asyncapi.checks.schemas;
+
+import org.sonar.check.Rule;
+import org.apiaddicts.apitools.dosonarapi.sslr.yaml.grammar.JsonNode;
+
+@Rule(key = AAR046AvroRecordDocCheck.CHECK_KEY)
+public class AAR046AvroRecordDocCheck extends AbstractAvroRecordCheck {
+    public static final String CHECK_KEY = "AAR046";
+    private static final String ERROR_KEY = "AAR046.error";
+
+    @Override
+    protected void visitAvroRecord(JsonNode node) {
+        JsonNode docNode = node.get("doc");
+        if (docNode.isMissing() || docNode.isNull()) {
+            addIssue(CHECK_KEY, translate(ERROR_KEY), node.key());
+        }
+    }
+}